calcappclient bugfix: "if" -> "else if". tcplikeserver: fixed %u printing of size_t by cast.
git-svn-id: https://svn.tdr.iem.uni-due.de:8443/svn/sctp/trunk/rsplib2@1576 516812f1-d3fd-0310-9ea4-9aededebcb87

+++ rsplib/tcplikeserver.cc
@@ -309,7 +309,7 @@ void TCPLikeServer::poolElement(const char*          programTitle,
             else {
                puts("off");
             }
-            printf("   Max Threads             = %u\n", maxThreads);
+            printf("   Max Threads             = %u\n", (unsigned int)maxThreads);
             puts("   Policy Settings");
             printf("      Policy Type          = %s\n", (policyName != NULL) ? policyName : "?");
             printf("      Load Degradation     = %1.3f [%%]\n", 100.0 * ((double)loadinfo->rli_load_degradation / (double)PPV_MAX_LOAD_DEGRADATION));
